requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.11.0
openpyxl>=3.1.0
